Warnings.

+++ tests/base16_test.c
@@ -32,7 +32,7 @@ int base16_test(void)
    for (idx = 0; idx < 2; idx++) {
       l1 = sizeof(out);
       DO(base16_encode(testin, sizeof(testin), out, &l1, idx));
-      COMPARE_TESTVECTOR(out, XSTRLEN(out), testout[idx], XSTRLEN(testout[idx]), "testout base16", idx);
+      COMPARE_TESTVECTOR(out, (unsigned long)XSTRLEN(out), testout[idx], (unsigned long)XSTRLEN(testout[idx]), "testout base16", idx);
       l2 = sizeof(tmp);
       DO(base16_decode(out, l1, tmp, &l2));
       COMPARE_TESTVECTOR(tmp, l2, testin, sizeof(testin), "testin base16", idx);
